Create an input field without the input tag.
The input tag can not be used if you want to have multiple active fields
at once. This is due to the fact that you must track which field the
user has selected in order to put the values in the correct space.

Tracking which filed they have selected is easy, but the problems arise
when they click buttons or if you want the user to be able to have no
field selected.

To select another field, some sort of event listener is needed to check
when the user clicks out of a input field, but this will happen when
they click on buttons.