Add a function that will check if there is a "missing" abs value.
As in it will only work if an abs value has been backspaced and will
"replace" it by changing the color back to black and moving the cursor
outside of the container instead of placing a new abs value.

Fixed some logical cursor bugs that resuleted from keepCursorPosition
being false when it shouldve been true.