Change the functionality of absValueThere and absValueBackspace
absValueThere will no longer place a bracket at the end no matter where
there cursor is. It will now only do that if the cursor is right before
the end. Instead, when a cursor is inside, it will move everything
after the cursor outside the absValue brackets and place the ending
bracket where the cursor is.

absValueBackspace now does the same, but in reverse. If there are
elements after the absValue container, it will encapsulate all
the remaining elements inside of it when the ending bracket is
deleted.