Add recursive function that gets cursor element.
This was done so that the cursor could be found
no matter how deeply nested it was inside of something.
As a result, hard cording to find the cursor positiong
is no longer needed.

Change any functions which had hard-coded cursor finding mechanisms.
-----------------------------------------------------------------------
Change the functionality of the absValue container.

When it is placed into the field, it will automatically
enclose everything of it's parent container starting from
the cursor position.

The container will now automatically delete itself when
the cursor is the first child, no matter how many children
are inside. However, the children are not deleted and are
moved to the outiside container.