Create functions relating to creating and deleting fields.
A field can't be created unless the currently selected one not is empty.
A field can now be deleted with backspace,
but only if that field is empty.

The function deleteField handles the deletion and subsequent
placing the cursor on the newly selected field.
The previous is always favored, unless it does not exist.
If there are no other fields nothing will happen.